Create useLocalStorage custom hook for stretch
