+++ client/src/components/CheckoutForm.js
@@ -1,4 +1,9 @@
 import React, { useState } from "react";
+import { useForm } from '../hooks/useForm'
+
+// This form should be handled by a "useForm" custom hook
+// Build out the logic needed for a form custom hook (see the useForm.js file)
+// and replace the necessary stateful logic from CheckoutForm with the hook
 
 const initialValue = {
   firstName: "",
@@ -9,22 +14,9 @@ const initialValue = {
   zip: "",
 };
 
-// This form should be handled by a "useForm" custom hook
-// Build out the logic needed for a form custom hook (see the useForm.js file)
-// and replace the necessary stateful logic from CheckoutForm with the hook
-
 const CheckoutForm = (props) => {
-  const [showSuccessMessage, setShowSuccessMessage] = useState(false);
-  const [values, setValues] = useState(initialValue);
-
-  const handleChanges = (e) => {
-    setValues({ ...values, [e.target.name]: e.target.value });
-  };
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    setShowSuccessMessage(true);
-  };
+  const [showSuccessMessage, setShowSuccessMessage] = useForm(false);
+  const [values, setValues] = useForm(initialValue);
 
   return (
     <>
